+++ src/test/java/com/example/fooddelivery/owner/service/OwnerServiceTest.java
@@ -14,7 +14,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.example.fooddelivery.common.exception.DuplicateException;
-import com.example.fooddelivery.common.exception.PasswordEncoder;
+import com.example.fooddelivery.common.PasswordEncoder;
 import com.example.fooddelivery.owner.domain.Owner;
 import com.example.fooddelivery.owner.dto.OwnerJoinReqDto;
 import com.example.fooddelivery.owner.repository.OwnerRepository;
